requirements.txt
streamlit>=1.28.0
pymongo>=4.5.0
python-dotenv>=1.0.0
Pillow>=10.2.0
plotly>=5.18.0
